pyproject.toml
[tool.poetry]
name = "reader-web-service"
version = "0.1.0"
description = ""
authors = ["Alexo <alexo@live.ru>"]
readme = "README.md"
packages = [{include = "reader_web_service"}]

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.2"
async-timeout = "^4.0.3"
selenium = "^4.18.1"
pyvirtualdisplay = "^3.0"
selenium-screenshot = "^2.1.0"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
telegraph = {extras = ["aio"], version = "^2.2.0"}
html-telegraph-poster = "^0.5.0"
pyroscope-io = "^0.8.5"
notifiers = "^1.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
